requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.4.0
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
coverage<=6.2
Django == 4.0.1
django-auto-prefetching == 0.2.11
django-cors-headers==3.11.0
django-filter==21.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt == 5.0.0
grpcio==1.43.0
httplib2==0.20.2
idna==3.3
iniconfig==1.1.1
mysql== 0.0.3
mysqlclient==2.1.0
packaging==21.3
phonenumbers==8.12.41
Pillow==9.0.0
pip==21.3.1
pluggy==1.0.0
proto-plus==1.19.8
protobuf==3.19.3
py==1.10.0
pydantic==1.9.0
PyJWT==2.3.0
pytest-cov==3.0.0
pytest==6.2.5
pytest-django==4.5.2
python-http-client==3.3.5
pytz==2021.3
requests== 2.27.1
sendgrid==6.9.4
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
toml==0.10.2
twilio==7.5.0
typing-extensions==4.0.1
uritemplate==4.1.1
mysql2postgresql== 0.8.5
django-postgresql== 0.0.3
pytest-postgresql ==4.1.0
py-postgresql ==1.2.2
psycopg ==3.0.8
psycopg2 == 2.9.3
psycopg-binary ==3.0.8
geopayment== 0.4.2
map== 1.3.0
iptools== 0.7.0
ip_intelligence ==1.1
IP2Locationpy ==8.0.1
ip2c== 0.1.2
ip2geotools== 0.1.6
ip2geotools-locator ==1.6.0
django-google-maps == 0.12.4
GDAL== 3.4.1

